feat: create agileplus-fixtures shared test crate
Extract test fixture code from agileplus-dashboard and agileplus-integration-tests
into a new shared crate for reuse across all test suites.

## Fixtures Extracted

### Modules:
- test_fixtures.rs: Core TestFixtures struct with pre-built feature test data
- payloads.rs: API payload builders (feature_create_payload, transition_payload, plane_webhook_payload)
- builders.rs: Fluent builder patterns for Features and WorkPackages
- dogfood.rs: AgilePlus dogfood features (4) + SpecKitty reference specs (33)

### Reusability:
- Seed data now available to dashboard, integration tests, CLI tests, contract tests
- Builder patterns enable flexible fixture composition
- Single source of truth for test data

### Impact:
- New shared fixtures crate: 999 LOC
- Reduced agileplus-dashboard/seed.rs: 541 LOC -> 9 LOC (re-exports)
- Reduced integration-tests/fixtures.rs: 146 LOC -> 13 LOC (re-exports)
- Net: 338 LOC + reusable library enabling Phase 1 LOC reduction goal

## Testing:
- All fixture builders include unit tests
- Dogfood data validated: 37 features, work packages, state transitions, labels
- Backward compatibility maintained via re-exports

Traceability: Phase 1 LOC Reduction, WP-Fixtures, WP19-T107
