+++ Cargo.toml
@@ -17,6 +17,7 @@ members = [
   "crates/agileplus-nats",
   "crates/agileplus-sync",
   "crates/agileplus-dashboard",
+  "crates/agileplus-fixtures",
   "crates/agileplus-github",
   "crates/agileplus-p2p",
   "crates/agileplus-integration-tests",
@@ -25,17 +26,17 @@ members = [
   "rust",
   "tests/bdd",
   "libs/nexus",
-  "libs/hexagonal-rs",
-  "libs/hexkit",
-  "libs/cipher",
-  "libs/gauge",
-  "libs/logger",
-  "libs/metrics",
-  "libs/tracing",
-  "libs/cli-framework",
-  "libs/config-core",
-  "libs/xdd-lib-rs",
-  "tools/forge",
+  # "libs/hexagonal-rs",
+  # "libs/hexkit",
+  # "libs/cipher",
+  # "libs/gauge",
+  # "libs/logger",
+  # "libs/metrics",
+  # "libs/tracing",
+  # "libs/cli-framework",
+  # "libs/config-core",
+  # "libs/xdd-lib-rs",
+  # "tools/forge",
 ]
 
 [workspace.package]
