+++ crates/agileplus-dashboard/Cargo.toml
@@ -12,6 +12,7 @@ dirs = ["../../templates"]
 agileplus-domain = { path = "../agileplus-domain" }
 agileplus-events = { path = "../agileplus-events" }
 agileplus-sqlite = { path = "../agileplus-sqlite" }
+agileplus-fixtures = { path = "../agileplus-fixtures" }
 askama = "0.12"
 axum = { version = "0.8", features = ["json", "macros"] }
 axum-extra = { version = "0.10", features = ["typed-header"] }
